+++ pages/api/auth/[...nextauth].js
@@ -28,8 +28,7 @@ const authOptions = {
             },
             async authorize(credentials, req) {
                 // Add logic here to look up the user from the credentials supplied
-                const user = await userController.findByEmail(credentials.email)
-
+                const [user,created] = await userController.findByEmail(credentials.email)
                 if (user) {
                     // Any object returned will be saved in `user` property of the JWT
                     return user
