+++ pages/home.js
@@ -1,24 +1,13 @@
-import { getSession } from "next-auth/react"
-import Profile from "./profile/index"
+import MainHeader from "../components/layout.js/main-header"
+export default function Home() {
+  return (
+    <>
+      <MainHeader />
+      <div>
+        Home page
+      </div>
+    </>
+  )
 
-export default function Home(props) {
-  return (<Profile currentUser={props.currentUser}/>)
 
-}
-
-export async function getServerSideProps(req, res) {
-  const session = await getSession(req)
-  console.log('session ', session)
-  if (!session) {
-    return {
-      redirect: {
-        permanent: false,
-        destination: `api/auth/signin?callbackUrl=http%3A%2F%2Flocalhost%3A3000%2F/`
-      }
-    }
-  }
-
-  return {
-    props: { currentUser: session?.user || null },
-  }
 }
