+++ pages/index.js
@@ -1,12 +1,28 @@
-function Home() {
+import { getSession } from "next-auth/react"
+import MainHeader from "../components/layout.js/main-header"
+export default function Home(props) {
   return (
-
-    <div>
-
-      <h1>Home Page</h1>
-    </div>
-
+    <>
+      <MainHeader currentUser={props.currentUser}/>
+    </>
   )
+
+
 }
 
-export default Home;
+export async function getServerSideProps(req, res) {
+  const session = await getSession(req)
+  console.log('sessions ', session)
+  if (!session) {
+    return {
+      redirect: {
+        permanent: false,
+        destination: `/home`
+      }
+    }
+  }
+
+  return {
+    props: { currentUser: session?.user || null },
+  }
+}
